+++ tt_metal/fabric/topology_solver.cpp
@@ -96,7 +96,7 @@ std::map<MeshId, AdjacencyGraph<FabricNodeId>> build_adjacency_graph_logical(con
         std::vector<FabricNodeId> adjacents;
         for (const auto& [neighbor_chip_id, edge] : adjacent_map) {
             // Skip self-connections
-            if (neighbor_chip_id == fabric_node_id.chip_id) {
+            if (static_cast<uint32_t>(neighbor_chip_id) == fabric_node_id.chip_id) {
                 continue;
             }
             for (size_t i = 0; i < edge.connected_chip_ids.size(); ++i) {
