+++ tt_metal/programming_examples/distributed/4_distributed_trace_and_events/distributed_trace_and_events.cpp
@@ -270,7 +270,7 @@ int main() {
     for (auto val : add_dst_vec) {
         pass &= (static_cast<float>(val) == workload_0_src0_val + workload_0_src1_val);
     }
-    for (int i = 0; i < mul_sub_dst_vec.size(); i++) {
+    for (size_t i = 0; i < mul_sub_dst_vec.size(); i++) {
         if (i < mul_sub_dst_vec.size() / 2) {
             pass &= (static_cast<float>(mul_sub_dst_vec[i]) == workload_1_src0_val * workload_1_src1_val);
         } else {
